requirements.txt
flask
pandas
flask_restful
requests
configparser
psycopg2-binary
sqlalchemy
